requirements.txt
et-xmlfile==1.0.1
jdcal==1.4.1
numpy==1.19.4
openpyxl==3.0.5
pandas==1.1.4
python-dateutil==2.8.1
pytz==2020.4
six==1.15.0
suds-jurko==0.6
xlrd==1.2.0
